optimize disambiguate.
The main change is that we don't iterate over the tagged file
for each tweet, but load it once (load another one only when the
tweet's file changes) into a dictionary, and later access the
values from there.

This moves the disambiguate from one of the biggest
time time users to "not measurable".